requirements.txt
httpx>=0.27.0
flask>=3.0.0
click>=8.1.0
rich>=13.0.0
peewee>=3.17.0
pandas>=2.2.0
plotly>=5.24.0
apscheduler>=3.10.0
cryptography>=43.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
